requirements.txt
torch
PyOpenGL
PyOpenGL_accelerate
urllib3
opencv-python
pillow
torchvision
torchaudio
numpy 
flask